Update 3/12/2015
@Author Charoenvit Sammanont

--Update on Grid System

+++ www/sync_components/sync.js
@@ -155,34 +155,58 @@ Sync.controller('SyncCtrl', ['$scope', '$location', 'AppDB', 'WebService', '_', 
                     $scope.pushDefectedResultData()
                 ]).then(function (results) {
 
+                    var _callBackDefected = JSON.parse(results[0].data);
+
+                    var _callBackDefectedResult = JSON.parse(results[1].data);
                     // call back from push Defected data and DefectedResults;
-                    //drop ALL TABLES 
-                    AppDB.dropAllTables();
 
-                    //create ALL TABLES
-                    AppDB.createAllTables();
+//                    alert(JSON.stringify(_callBackDefected));
+//                    
+//                    alert(JSON.stringify(_callBackDefectedResult));
 
-                    //sync data from service
-                    AppDB.syncData().then(function (result) {
+                    if (_callBackDefected.ErrorCode === _success
+                            && _callBackDefectedResult.ErrorCode === _success)
+                    {
+                        //drop ALL TABLES 
+                        AppDB.dropAllTables();
 
-                        toastr.success('Webservice Called Complete!', 'Information', {
-                            timeOut: 10000
-                        });
+                        //create ALL TABLES
+                        AppDB.createAllTables();
 
-                        //insert data into tables 
-                        AppDB.insertData(result).then(function () {
+                        //sync data from service
+                        AppDB.syncData().then(function (result) {
 
-                            cfpLoadingBar.complete();
-
-                            toastr.success('Synced Complete!', 'Information', {
-                                timeOut: 5000
+                            toastr.success('Webservice Called Complete!', 'Information', {
+                                timeOut: 10000
                             });
 
-                            var _loginViewPath = '/loginView';
-                            //change page to login view page
-                            $location.path(_loginViewPath).replace();
+                            //insert data into tables 
+                            AppDB.insertData(result).then(function () {
 
-                        }, function (res) {
+                                cfpLoadingBar.complete();
+
+                                toastr.success('Synced Complete!', 'Information', {
+                                    timeOut: 5000
+                                });
+
+                                var _loginViewPath = '/loginView';
+                                //change page to login view page
+                                $location.path(_loginViewPath).replace();
+
+                            }, function (res) {
+
+                                cfpLoadingBar.complete();
+
+                                //drop ALL TABLES 
+                                AppDB.dropSetupTable();
+                                AppDB.dropAllTables();
+
+                                toastr.error('Synced Falied at TABLE : ' + res.Table, 'Error', {
+                                    timeOut: 5000
+                                });
+                            });
+
+                        }, function (result) {
 
                             cfpLoadingBar.complete();
 
@@ -190,23 +214,11 @@ Sync.controller('SyncCtrl', ['$scope', '$location', 'AppDB', 'WebService', '_', 
                             AppDB.dropSetupTable();
                             AppDB.dropAllTables();
 
-                            toastr.error('Synced Falied at TABLE : ' + res.Table, 'Error', {
+                            toastr.error('Could not connect to ' + result.config.url, 'Error', {
                                 timeOut: 5000
                             });
                         });
-
-                    }, function (result) {
-
-                        cfpLoadingBar.complete();
-
-                        //drop ALL TABLES 
-                        AppDB.dropSetupTable();
-                        AppDB.dropAllTables();
-
-                        toastr.error('Could not connect to ' + result.config.url, 'Error', {
-                            timeOut: 5000
-                        });
-                    });
+                    }
                 }, function (results) {
 
                     cfpLoadingBar.complete();
